finishing authentication
