+++ App.js
@@ -1,4 +1,6 @@
 import { createAppContainer, createSwitchNavigator } from 'react-navigation';
+import React from 'react';
+import { SafeAreaProvider } from 'react-native-safe-area-context';
 import firebase from 'firebase';
 import { firebaseConfig } from './config/firebase.js';
 import AuthNavigator from './navigation/AuthNavigator';
@@ -6,8 +8,7 @@ import HomeScreen from './components/HomeScreen.js';
 
 firebase.initializeApp(firebaseConfig);
 
-export default createAppContainer(
-
+const AppContainer = createAppContainer(
     createSwitchNavigator(
         {
             Auth: AuthNavigator,
@@ -18,3 +19,9 @@ export default createAppContainer(
         }
     )
 );
+
+const Plants = () => {
+    return (<SafeAreaProvider><AppContainer></AppContainer></SafeAreaProvider>)
+}
+
+export default Plants;
