+++ components/LoadingScreen.js
@@ -7,7 +7,7 @@ class LoadingScreen extends React.Component {
             if (user) {
             this.props.navigation.navigate('App');
             } else {
-            this.props.navigation.navigate('SignUp');
+            this.props.navigation.navigate('SignIn');
             }
         });
     }
