+++ config/facebook.js
@@ -1,0 +1,5 @@
+import {
+    FACEBOOK_APP_ID
+} from '@env'
+
+export const facebookAppId = FACEBOOK_APP_ID
