+++ config/google.js
@@ -1,0 +1,7 @@
+import {
+    ANDROID_CLIENT_ID,
+    IOS_CLIENT_ID
+} from '@env'
+
+export const androidClientId = ANDROID_CLIENT_ID
+export const IOSClientId = IOS_CLIENT_ID
