+++ package.json
@@ -21,7 +21,8 @@
     "react-native-dotenv": "^2.4.3",
     "react-native-gesture-handler": "^1.9.0",
     "react-native-web": "~0.13.12",
-    "react-navigation": "^4.4.3"
+    "react-navigation": "^4.4.3",
+    "react-native-safe-area-context": "3.1.4"
   },
   "devDependencies": {
     "@babel/core": "~7.9.0"
